Extract utils to separate utils module.
This will allow static utility code to be used by things that don't need to depend on the entirety of the core module
