+++ recaf-assembler/build.gradle
@@ -8,12 +8,14 @@ repositories {
 }
 
 dependencies {
-    def antlrVersion = '4.9.2'
+    api project(':recaf-utils')
 
+    def antlrVersion = '4.9.2'
     antlr "org.antlr:antlr4:$antlrVersion"
-    implementation "org.antlr:antlr4-runtime:$antlrVersion"
-    implementation asm
-    implementation asm_tree
+    api "org.antlr:antlr4-runtime:$antlrVersion"
+
+    api asm
+    api asm_tree
 
     // Testing
     testImplementation junit_api
