+++ recaf-assembler/src/main/java/me/coley/recaf/assemble/parser/BytecodeVisitorImpl.java
@@ -5,7 +5,6 @@ import me.coley.recaf.assemble.ast.arch.*;
 import me.coley.recaf.assemble.ast.insn.*;
 import me.coley.recaf.assemble.ast.meta.Comment;
 import me.coley.recaf.assemble.ast.meta.Label;
-import me.coley.recaf.assemble.util.ParserException;
 import org.antlr.v4.runtime.ParserRuleContext;
 import org.antlr.v4.runtime.Token;
 import org.antlr.v4.runtime.tree.ParseTree;
