+++ recaf-assembler/src/main/java/me/coley/recaf/assemble/parser/ParserException.java
@@ -1,4 +1,4 @@
-package me.coley.recaf.assemble.util;
+package me.coley.recaf.assemble.parser;
 
 import org.antlr.v4.runtime.tree.ParseTree;
