--- recaf-assembler/src/main/java/me/coley/recaf/assemble/util/Placeholder.java
@@ -1,10 +1,0 @@
-package me.coley.recaf.assemble.util;
-
-public class Placeholder {
-	public static String escape(String text) {
-		// TODO: Handle string escapes
-		//  - escape util is in core, so we can't access it here...
-		//  - one fix would be to move it to this module and have core require this module
-		return text;
-	}
-}
