+++ recaf-core/build.gradle
@@ -5,6 +5,8 @@ plugins {
 apply plugin: 'java-library'
 
 dependencies {
+    api project(':recaf-utils')
+
     api asm
     api asm_analysis
     api asm_commons
@@ -20,9 +22,8 @@ dependencies {
     api jphantom
     api logback_classic
     api picocli
-    api guava
     api gson
-    api regex
+
     api smali
 
     // Java 9+ compatibility
