+++ recaf-utils/build.gradle
@@ -1,0 +1,22 @@
+apply plugin: 'java-library'
+
+group 'me.coley'
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    api asm
+    api asm_tree
+    api guava
+    api logback_classic
+    api regex
+
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.0'
+    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.7.0'
+}
+
+test {
+    useJUnitPlatform()
+}
