+++ recaf-utils/src/main/java/me/coley/recaf/util/EscapeUtil.java
