+++ recaf-utils/src/main/java/me/coley/recaf/util/OperatingSystem.java
