+++ recaf-utils/src/main/java/me/coley/recaf/util/ShortcutUtil.java
