+++ recaf-utils/src/main/java/me/coley/recaf/util/Types.java
@@ -1,0 +1,9 @@
+package me.coley.recaf.util;
+
+/**
+ * A wrapper around {@link org.objectweb.asm.Type}
+ */
+public class Types {
+	// TODO: Create wrapper calls for common Type functions that throw silent exceptions like IllegalArgumentException
+	//  - getTypeInternal
+}
