+++ recaf-utils/src/main/java/me/coley/recaf/util/logging/Logging.java
